hpがsize_tだとオーバーフローするのでintにした

+++ AnnihilationCharacter_Siv3D/AnnihilationCharacter_Siv3D/Game/inc/Boss.hpp
@@ -15,7 +15,7 @@ namespace game {
 	class Boss {
 		std::unique_ptr<BossBehavior> m_bossBehavior;
 		const Level m_level;
-		size_t m_hp;
+		int m_hp;
 		std::unique_ptr<BulletListCreator> m_bulletListCreator;
 		std::unique_ptr<elipmocframework::FontObject> m_fontObject;
 		std::unique_ptr<CollisionCircle> m_collision;
@@ -25,8 +25,8 @@ namespace game {
 		static constexpr siv::Vec2 BOSS_POS = { 570, 160 };
 		Boss(Level level,const siv::Vec2& playerRefPos, BulletListCreator&);
 		~Boss();
-		size_t GetHp() { return m_hp; }
-		size_t SetHp(size_t hp) { return m_hp = hp; }
+		int GetHp() { return m_hp; }
+		int SetHp(int hp) { return m_hp = hp; }
 		void SetBossBehavior();
 		void Update();
 		void Draw()const;
